# src/main/resources/static/swagger-config.yml
swagger: '2.0'
info:
  version: '1.0'
  title: Strijkaterlier API
  contact:
    name: Lucas Vda
host: 'localhost:8081'
basePath: /
...
paths:
  /api/baskets/:
    get:
      tags:
        - basket-management
      summary: 'get all baskets '
      description: 'Gets all baskets from database '
      operationId: getBaskets
      produces:
        - application/json
      parameters:
        - name: page
          in: query
          description: page
          required: false
          type: integer
          default: 1
          format: int32
      responses:
        '200':
          description: '{baskets}'
          schema:
            type: array
            items:
              $ref: '#/definitions/BasketDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal server error
      deprecated: false
    post:
      tags:
        - basket-management
      summary: adds basket
      description: 'adds a basket to the database '
      operationId: addBasket
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: basketDto
          description: basketDto
          required: true
          schema:
            $ref: '#/definitions/BasketDto_1'
      responses:
        '200':
          description: '{basket}'
          schema:
            $ref: '#/definitions/BasketDto'
        '201':
          description: Created
        '400':
          description: Invalid request
          schema:
            $ref: '#/definitions/Errors'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '409':
          description: 'basket already exists {basketId}'
        '500':
          description: Internal server error
      deprecated: false
  /api/baskets/active:
    get:
      tags:
        - basket-management
      summary: get active baskets
      description: Gets all active baskets from database
      operationId: getActiveBaskets
      produces:
        - application/json
      responses:
        '200':
          description: '{baskets}'
          schema:
            type: array
            items:
              $ref: '#/definitions/BasketDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal server error
      deprecated: false
  '/api/baskets/active/{startDate}':
    get:
      tags:
        - basket-management
      summary: get active baskets by startDate
      description: Gets all active baskets by startdate from database
      operationId: getActiveBasketsByStartDate
      produces:
        - application/json
      parameters:
        - name: startDate
          in: path
          description: startDate
          required: true
          type: string
          format: date-time
      responses:
        '200':
          description: '{baskets}'
          schema:
            type: array
            items:
              $ref: '#/definitions/BasketDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal server error
      deprecated: false
  '/api/baskets/{basketId}':
    get:
      tags:
        - basket-management
      summary: 'get basket by id '
      description: Gets baskets by id from database
      operationId: getBasketsById
      produces:
        - application/json
      parameters:
        - name: basketId
          in: path
          description: basketId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: '{basket}'
          schema:
            $ref: '#/definitions/BasketDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: 'no basket found with name {itemName}'
        '500':
          description: Internal server error
      deprecated: false
    put:
      tags:
        - basket-management
      summary: updates basket
      description: 'updates a basket to the database '
      operationId: updateBasket
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: basketDto
          description: basketDto
          required: true
          schema:
            $ref: '#/definitions/BasketDto_1'
        - name: basketId
          in: path
          description: basketId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: '{basket}'
          schema:
            $ref: '#/definitions/BasketDto'
        '201':
          description: Created
        '400':
          description: Invalid request
          schema:
            $ref: '#/definitions/Errors'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: 'basket doesnt exists {basketId}'
        '500':
          description: Internal server error
      deprecated: false
    delete:
      tags:
        - basket-management
      summary: delete basket
      description: 'delets a basket to the database '
      operationId: deleteBasket
      produces:
        - '*/*'
      parameters:
        - name: basketId
          in: path
          description: basketId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: '{basket}'
        '204':
          description: No Content
        '400':
          description: Invalid request
          schema:
            $ref: '#/definitions/Errors'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: 'basket doesnt exists {basketId}'
        '500':
          description: Internal server error
      deprecated: false
    patch:
      tags:
        - basket-management
      summary: set items in basket
      description: 'sets the items of a basket to the database '
      operationId: setItems
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: basketId
          in: path
          description: basketId
          required: true
          type: integer
          format: int64
        - in: body
          name: itemDtos
          description: itemDtos
          required: true
          schema:
            type: array
            items:
              $ref: '#/definitions/ItemDto_1'
      responses:
        '200':
          description: '{basket}'
        '204':
          description: No Content
        '400':
          description: Invalid request
          schema:
            $ref: '#/definitions/Errors'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: 'basket doesnt exists {basketId}'
        '500':
          description: Internal server error
      deprecated: false
  '/api/baskets/{basketId}/items':
    post:
      tags:
        - basket-management
      summary: adds item in basket
      description: 'adds  item of a basket to the database '
      operationId: addItem
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: basketId
          in: path
          description: basketId
          required: true
          type: integer
          format: int64
        - in: body
          name: itemDto
          description: itemDto
          required: true
          schema:
            $ref: '#/definitions/ItemDto_1'
      responses:
        '200':
          description: '{basket}'
          schema:
            $ref: '#/definitions/BasketDto'
        '201':
          description: Created
        '400':
          description: Invalid request
          schema:
            $ref: '#/definitions/Errors'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: 'basket doesnt exists {basketId}'
        '500':
          description: Internal server error
      deprecated: false
...
  '/api/customer/basket/{basketId}':
    put:
      tags:
        - customer-management
      summary: updates a basket from a customer
      description: update a basket of a customer in the database
      operationId: updateBasketFromCustomer
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: basket
          description: basket
          required: true
          schema:
            $ref: '#/definitions/Basket'
        - name: basketId
          in: path
          description: basketId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: basket is successfully updated
        '201':
          description: Created
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/Errors'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: customer is not found
          schema:
            $ref: '#/definitions/Errors'
        '500':
          description: Internal server error
      deprecated: false
...
definitions:
  Basket:
    type: object
    properties:
      active:
        type: boolean
      cash:
        type: boolean
      customer:
        $ref: '#/definitions/Customer'
      doneDateTime:
        type: string
      id:
        type: integer
        format: int64
      items:
        type: array
        items:
          $ref: '#/definitions/Item'
      price:
        type: number
        format: double
      startDateTime:
        type: string
    title: Basket
  BasketDto:
    type: object
    required:
      - active
      - cash
      - customerDto
      - doneDateTime
      - itemDtos
      - startDateTime
    properties:
      customerDto:
        example:
          customerDto: null
        description: Customer of the basket
        $ref: '#/definitions/CustomerDto'
      id:
        type: integer
        format: int64
        example: 3
        description: Id of the Basket
      active:
        type: boolean
        example: true
        description: Is basket active
      startDateTime:
        type: string
        example: '2/07/2020T18:57:20'
        description: Startdate of the Basket
      doneDateTime:
        type: string
        example: '2/07/2020T18:57:20'
        description: Date when Basket is done
      itemDtos:
        type: array
        example:
          items: null
        description: items of the Basket
        items:
          $ref: '#/definitions/ItemDto'
      price:
        type: number
        format: double
        example: 20
        description: Startdate of the Basket
      cash:
        type: boolean
        example: false
        description: payment method of Basket
    title: BasketDto
    description: Request for Basket
  BasketDto_1:
    type: object
    required:
      - active
      - cash
      - customerDto
      - doneDateTime
      - itemDtos
      - startDateTime
    properties:
      customerDto:
        example:
          customerDto: null
        description: Customer of the basket
        $ref: '#/definitions/CustomerDto_1'
      id:
        type: integer
        format: int64
        example: 3
        description: Id of the Basket
      active:
        type: boolean
        example: true
        description: Is basket active
      startDateTime:
        type: string
        example: '2/07/2020T18:57:20'
        description: Startdate of the Basket
      doneDateTime:
        type: string
        example: '2/07/2020T18:57:20'
        description: Date when Basket is done
      itemDtos:
        type: array
        example:
          items: null
        description: items of the Basket
        items:
          $ref: '#/definitions/ItemDto_1'
      price:
        type: number
        format: double
        example: 20
        description: Startdate of the Basket
      cash:
        type: boolean
        example: false
        description: payment method of Basket
    title: BasketDto_1
    description: Request for Basket
  Customer:
    type: object
    properties:
      baskets:
        type: array
        items:
          $ref: '#/definitions/Basket'
      emailaddress:
        type: string
      firstName:
        type: string
      id:
        type: integer
        format: int64
      lastName:
        type: string
      minutesLeft:
        type: number
        format: double
      note:
        type: string
      phoneNumber:
        type: string
    title: Customer
  CustomerDto:
    type: object
    required:
      - emailaddress
      - firstName
      - lastName
      - minutesLeft
      - phoneNumber
    properties:
      id:
        type: integer
        format: int64
        example: 3
        description: Id of the Customer
      emailaddress:
        type: string
        example: 3
        description: Email of the customer
      firstName:
        type: string
        example: 3
        description: Firstname of the customer
      lastName:
        type: string
        example: 3
        description: Lastname of the customer
      phoneNumber:
        type: string
        example: 3
        description: Phonenumber of the customer
      basketDtos:
        type: array
        example: 3
        description: Baskets of the customer
        items:
          $ref: '#/definitions/BasketDto'
      minutesLeft:
        type: number
        format: double
        example: 3
        description: MinutesLeft of the customer
      note:
        type: string
        example: 3
        description: Notes of the customer
    title: CustomerDto
    description: CustomersDto
  CustomerDto_1:
    type: object
    required:
      - firstName
      - lastName
      - minutesLeft
      - phoneNumber
    properties:
      emailAddress:
        type: string
      emailaddress:
        type: string
      id:
        type: integer
        format: int64
        example: 3
        description: Id of the Customer
      firstName:
        type: string
        example: 3
        description: Firstname of the customer
      lastName:
        type: string
        example: 3
        description: Lastname of the customer
      phoneNumber:
        type: string
        example: 3
        description: Phonenumber of the customer
      basketDtos:
        type: array
        example: 3
        description: Baskets of the customer
        items:
          $ref: '#/definitions/BasketDto_1'
      minutesLeft:
        type: number
        format: double
        example: 3
        description: MinutesLeft of the customer
      note:
        type: string
        example: 3
        description: Notes of the customer
    title: CustomerDto_1
    description: CustomersDto
  ErrorInfo:
    type: object
    required:
      - description
      - message
    properties:
      message:
        type: string
        example: Missing field userName
        description: Error Message
      description:
        type: string
        example: userName field is required for registering a new user
        description: Description of the problem
      moreInfo:
        type: string
        example: Please make sure that the requestbody contains the field userName
        description: More information on the problem and or recommendations to resolve the error
    title: ErrorInfo
    description: Error detail
  Errors:
    type: object
    required:
      - errors
    properties:
      errors:
        type: array
        description: List of ErrorInfo
        items:
          $ref: '#/definitions/ErrorInfo'
    title: Errors
    description: List of Errors that occurred
  Item:
    type: object
    properties:
      baskets:
        type: array
        items:
          $ref: '#/definitions/Basket'
      id:
        type: integer
        format: int64
      itemName:
        type: string
      minutes:
        type: integer
        format: int32
      price:
        type: number
        format: double
    title: Item
  ItemDto:
    type: object
    required:
      - itemName
      - minutes
      - price
    properties:
      basketDtos:
        type: array
        items:
          $ref: '#/definitions/BasketDto'
      id:
        type: integer
        format: int64
        example: 3
        description: Id of the item
      itemName:
        type: string
        example: Rok
        description: Name of the item
      price:
        type: number
        format: double
        example: 3.5
        description: Price of the item
      minutes:
        type: integer
        format: int32
        example: 10
        description: Minutes of the item
    title: ItemDto
    description: Request for Items
  ItemDto_1:
    type: object
    required:
      - itemName
      - minutes
      - price
    properties:
      basketDtos:
        type: array
        items:
          $ref: '#/definitions/BasketDto_1'
      id:
        type: integer
        format: int64
        example: 3
        description: Id of the item
      itemName:
        type: string
        example: Rok
        description: Name of the item
      price:
        type: number
        format: double
        example: 3.5
        description: Price of the item
      minutes:
        type: integer
        format: int32
        example: 10
        description: Minutes of the item
    title: ItemDto_1
    description: Request for Items
...